# backend/app/domain/enums.py
from enum import StrEnum


class UserRole(StrEnum):
    EXECUTIVE = "EXECUTIVE"
    MANAGER = "MANAGER"


class Category(StrEnum):
    BLOQUEO_TARJETA = "BLOQUEO_TARJETA"
    REPORTE_FRAUDE = "REPORTE_FRAUDE"
    CONSULTA_GENERAL = "CONSULTA_GENERAL"
    SOLICITUD_CREDITO = "SOLICITUD_CREDITO"
    BANCA_DIGITAL = "BANCA_DIGITAL"


class Priority(StrEnum):
    CRITICO = "CRITICO"
    ALTO = "ALTO"
    MEDIO = "MEDIO"
    BAJO = "BAJO"


class ConsultationLevel(StrEnum):
    GENERAL = "GENERAL"
    PERSONALIZADA = "PERSONALIZADA"
    SENSIBLE = "SENSIBLE"


class IdentificationStatus(StrEnum):
    ANONIMO = "ANONIMO"
    PENDIENTE = "PENDIENTE"
    IDENTIFICADO = "IDENTIFICADO"
    FALLIDO = "FALLIDO"
    VERIFICACION_MANUAL = "VERIFICACION_MANUAL"


class SessionStatus(StrEnum):
    CREATED = "CREATED"
    LISTENING = "LISTENING"
    NEEDS_CLARIFICATION = "NEEDS_CLARIFICATION"
    AWAITING_CONFIRMATION = "AWAITING_CONFIRMATION"
    AWAITING_IDENTIFICATION = "AWAITING_IDENTIFICATION"
    ORCHESTRATING = "ORCHESTRATING"
    RESOLVED_AUTOMATIC = "RESOLVED_AUTOMATIC"
    ASSIGNED = "ASSIGNED"
    FAILED = "FAILED"


class CaseStatus(StrEnum):
    CREATED = "CREATED"
    CLASSIFIED = "CLASSIFIED"
    ORCHESTRATING = "ORCHESTRATING"
    RESOLVED = "RESOLVED"
    QUEUED = "QUEUED"
    ASSIGNED = "ASSIGNED"
    CLOSED = "CLOSED"


class TicketStatus(StrEnum):
    PENDIENTE = "PENDIENTE"
    EN_ATENCION = "EN_ATENCION"
    CERRADO = "CERRADO"


class ResolutionOutcome(StrEnum):
    RESUELTO = "RESUELTO"
    DERIVADO = "DERIVADO"
    PENDIENTE_DOCUMENTACION = "PENDIENTE_DOCUMENTACION"
    CLIENTE_DESISTIO = "CLIENTE_DESISTIO"
    NO_RESUELTO = "NO_RESUELTO"


class ConversationRole(StrEnum):
    CUSTOMER = "CUSTOMER"
    ASSISTANT = "ASSISTANT"


class ExecutiveStatus(StrEnum):
    DISPONIBLE = "DISPONIBLE"
    OCUPADO = "OCUPADO"
    INACTIVO = "INACTIVO"


class ResolutionType(StrEnum):
    AUTOMATIC = "AUTOMATIC"
    HUMAN = "HUMAN"


class GroundingStatus(StrEnum):
    GROUNDED = "GROUNDED"
    NO_EVIDENCE = "NO_EVIDENCE"
    NOT_APPLICABLE = "NOT_APPLICABLE"


class KnowledgeSourceType(StrEnum):
    OFFICIAL = "OFFICIAL"
    REGULATORY = "REGULATORY"
    INTERNAL = "INTERNAL"
    HYBRID = "HYBRID"


class KnowledgeIndexStatus(StrEnum):
    PENDING = "PENDING"
    INDEXING = "INDEXING"
    READY = "READY"
    FAILED = "FAILED"
    ARCHIVED = "ARCHIVED"


class KnowledgeJobOperation(StrEnum):
    CREATE = "CREATE"
    VERSION = "VERSION"
    REINDEX = "REINDEX"


class KnowledgeJobStatus(StrEnum):
    QUEUED = "QUEUED"
    RUNNING = "RUNNING"
    SUCCEEDED = "SUCCEEDED"
    FAILED = "FAILED"