# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "Category": {
        "enum": [
          "BLOQUEO_TARJETA",
          "REPORTE_FRAUDE",
          "CONSULTA_GENERAL",
          "SOLICITUD_CREDITO",
          "BANCA_DIGITAL"
        ],
        "title": "Category",
        "type": "string"
      },
...
      "ConsultationLevel": {
        "enum": [
          "GENERAL",
          "PERSONALIZADA",
          "SENSIBLE"
        ],
        "title": "ConsultationLevel",
        "type": "string"
      },
...
      "ExecutiveAssignment": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "window_number": {
            "title": "Window Number",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "title",
          "window_number"
        ],
        "title": "ExecutiveAssignment",
        "type": "object"
      },
...
      "FlowResult": {
        "properties": {
          "citations": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeCitation"
            },
            "title": "Citations",
            "type": "array"
          },
          "customer_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Summary"
          },
          "executive": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecutiveAssignment"
              },
              {
                "type": "null"
              }
            ]
          },
          "grounding_status": {
            "$ref": "#/components/schemas/GroundingStatus",
            "default": "NOT_APPLICABLE"
          },
          "identification_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IdentificationStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "next_action": {
            "enum": [
              "CAPTURE",
              "IDENTIFY",
              "COMPLETE"
            ],
            "title": "Next Action",
            "type": "string"
          },
          "priority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Priority"
              },
              {
                "type": "null"
              }
            ]
          },
          "requirement_id": {
            "format": "uuid",
            "title": "Requirement Id",
            "type": "string"
          },
          "resolution_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResolutionType"
              },
              {
                "type": "null"
              }
            ]
          },
          "response": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "speech_text": {
            "title": "Speech Text",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "ticket": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TicketResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "tracking_information": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracking Information"
          }
        },
        "required": [
          "session_id",
          "requirement_id",
          "status",
          "next_action",
          "speech_text"
        ],
        "title": "FlowResult",
        "type": "object"
      },
      "GroundingStatus": {
        "enum": [
          "GROUNDED",
          "NO_EVIDENCE",
          "NOT_APPLICABLE"
        ],
        "title": "GroundingStatus",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IdentificationStatus": {
        "enum": [
          "ANONIMO",
          "PENDIENTE",
          "IDENTIFICADO",
          "FALLIDO",
          "VERIFICACION_MANUAL"
        ],
        "title": "IdentificationStatus",
        "type": "string"
      },
...
      "KnowledgeCitation": {
        "properties": {
          "chunk_id": {
            "format": "uuid",
            "title": "Chunk Id",
            "type": "string"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "score": {
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Score",
            "type": "number"
          },
          "section": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "chunk_id",
          "title",
          "page",
          "score"
        ],
        "title": "KnowledgeCitation",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "Priority": {
        "enum": [
          "CRITICO",
          "ALTO",
          "MEDIO",
          "BAJO"
        ],
        "title": "Priority",
        "type": "string"
      },
...
      "ResolutionType": {
        "enum": [
          "AUTOMATIC",
          "HUMAN"
        ],
        "title": "ResolutionType",
        "type": "string"
      },
      "SessionCreateRequest": {
        "properties": {
          "preferential_attention": {
            "default": false,
            "title": "Preferential Attention",
            "type": "boolean"
          }
        },
        "title": "SessionCreateRequest",
        "type": "object"
      },
      "SessionCreatedResponse": {
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "session_token": {
            "title": "Session Token",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          }
        },
        "required": [
          "session_id",
          "session_token",
          "status",
          "expires_at"
        ],
        "title": "SessionCreatedResponse",
        "type": "object"
      },
      "SessionStatus": {
        "enum": [
          "CREATED",
          "LISTENING",
          "NEEDS_CLARIFICATION",
          "AWAITING_CONFIRMATION",
          "AWAITING_IDENTIFICATION",
          "ORCHESTRATING",
          "RESOLVED_AUTOMATIC",
          "ASSIGNED",
          "FAILED"
        ],
        "title": "SessionStatus",
        "type": "string"
      },
      "SessionStatusResponse": {
        "properties": {
          "analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TurnAnalysisResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "final_response": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Response"
          },
          "resolution_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResolutionType"
              },
              {
                "type": "null"
              }
            ]
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FlowResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          }
        },
        "required": [
          "session_id",
          "status"
        ],
        "title": "SessionStatusResponse",
        "type": "object"
      },
...
      "TicketResult": {
        "properties": {
          "estimated_wait_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Wait Minutes"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "number": {
            "title": "Number",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/TicketStatus"
          }
        },
        "required": [
          "id",
          "number",
          "status"
        ],
        "title": "TicketResult",
        "type": "object"
      },
      "TicketStatus": {
        "enum": [
          "PENDIENTE",
          "EN_ATENCION",
          "CERRADO"
        ],
        "title": "TicketStatus",
        "type": "string"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "token_type": {
            "const": "bearer",
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserSummary"
          }
        },
        "required": [
          "access_token",
          "expires_in",
          "user"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
...
      "TurnAnalysisResponse": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "clarification_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clarification Question"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "consultation_level": {
            "$ref": "#/components/schemas/ConsultationLevel"
          },
          "customer_summary": {
            "title": "Customer Summary",
            "type": "string"
          },
          "next_action": {
            "enum": [
              "CLARIFY",
              "CONFIRM"
            ],
            "title": "Next Action",
            "type": "string"
          },
          "pii_types": {
            "items": {
              "type": "string"
            },
            "title": "Pii Types",
            "type": "array"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "requirement_id": {
            "format": "uuid",
            "title": "Requirement Id",
            "type": "string"
          },
          "speech_text": {
            "title": "Speech Text",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "requirement_id",
          "status",
          "summary",
          "customer_summary",
          "category",
          "priority",
          "consultation_level",
          "confidence",
          "next_action",
          "speech_text"
        ],
        "title": "TurnAnalysisResponse",
        "type": "object"
      },
...
      "UserRole": {
        "enum": [
          "EXECUTIVE",
          "MANAGER"
        ],
        "title": "UserRole",
        "type": "string"
      },
      "UserSummary": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "executive_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Executive Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        },
        "required": [
          "id",
          "email",
          "role"
        ],
        "title": "UserSummary",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API del sistema de orquestacion de atencion bancaria",
    "title": "Sistema de Orquestacion Bancaria",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "Autenticacion"
        ]
      }
    },
...
    "/api/v1/kiosk/sessions": {
      "post": {
        "operationId": "create_session_api_v1_kiosk_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCreatedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Session",
        "tags": [
          "Kiosco"
        ]
      }
    },
    "/api/v1/kiosk/sessions/{session_id}": {
      "get": {
        "operationId": "get_session_status_api_v1_kiosk_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Session-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session Status",
        "tags": [
          "Kiosco"
        ]
      }
    },
...
  }
}